refactor: ComedyPerformanceCalculator에 amountFor 메서드 오버라이드
